feat: implement tool processing logic in markdown plugin
- Added a new nodeProcessor module to handle tool use tags in markdown content.
- Implemented functions to process raw, paragraph, and code nodes for tool tags.
- Introduced utility functions for extracting tool attributes and decoding URLs.
- Ensured adherence to performance optimization practices and safe property access throughout the implementation.
